% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datathreshfiltr-package.R
\docType{package}
\name{datathreshfiltr-package}
\alias{datathreshfiltr-package}
\title{datathreshfiltr: A package for filtering CE-MS metabolomics peak area data}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/liamthepineapple/datathreshfiltr}
  \item Report bugs at \url{https://github.com/liamthepineapple/datathreshfiltr/issues}
}

}
\author{
\strong{Maintainer}: Liam Surry \email{surryl@mcmaster.ca} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
