% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datathreshfiltr.R
\name{datathreshfiltr}
\alias{datathreshfiltr}
\title{Filters your and replace values below}
\usage{
datathreshfiltr(file_name, threshold)
}
\arguments{
\item{file_name}{The name of your file in .csv format (character)}

\item{threshold}{The threshold value you want to filter your data at. (numeric)}
}
\value{
A datafile with a name of your choosing from your input \code{file_name} that has values below \code{threshold} replaced with NA
}
\description{
Filters your and replace values below
}
\examples{
data_file <- datathreshfiltr("file_name.csv", threshold = 15000)


}
