% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datathreshfiltr.R
\docType{data}
\name{metabolite_peak_areas}
\alias{metabolite_peak_areas}
\title{Peak area data from CE-MS metabolomics data}
\format{
A dataframe consiting of 482 rows and 82 columns
\describe {
\item{file.name}{name of sample file as numeric}
\item{peak.number}{peak number position in sample run between 1-13 as numeric}
\item{metabolite}{metabolites listed by individual mass as character}
}
}
\source{
Liam Surry and Zachery Krozen, Britz-Mckibbin laboratory, McMaster University, Department of Chemistry and Chemical Biology
}
\usage{
metabolite_peak_areas
}
\description{
A dataset containing peak area data from metabolomic drugs of abuse screening after being automatically preprocessed by PeakMeister.
}
\keyword{datasets}
